+++ src/components/signin-page/SigninForm.vue
@@ -1,73 +1,80 @@
 <script lang="ts" setup>
 import { ref } from 'vue'
-import type { FormInstance } from 'element-plus'
-import { useRouter, type Router } from 'vue-router'
-import { useAuthStore } from '@/stores/auth'
-const authStore = useAuthStore()
+import type { FormRules,FormInstance } from 'element-plus'
+// import { useRouter, type Router } from 'vue-router'
+// import { useAuthStore } from '@/stores/auth'
+// const authStore = useAuthStore()
 const ruleFormRef = ref<FormInstance>()
-const isLoading = ref<Boolean>(false)
-type SigninInputType = {
-  account: string
-  password: string
-}
-const router: Router = useRouter()
-const signinInput = ref<SigninInputType>({
-  account: '',
-  password: ''
-})
-const registerRules = ref({
-  account: [
-    {
-      type: 'email',
-      required: true,
-      message: '信箱格式不相符',
-      trigger: ['blur', 'change']
-    }
-  ],
-  password: [
-    { min: 2, max: 30, message: '長度介於6到30之間', trigger: 'blur' },
-    { required: true, message: '必填', trigger: 'blur' }
-  ]
-})
-const handleRegister = async (formEl: FormInstance | undefined) => {
-  if (!formEl) return
-  await formEl.validate(async (valid, fields) => {
-    if (valid) {
-      await fetchSignin(signinInput.value)
-      // console.log('發送登入API')
-    } else {
-      // console.log('error submit!', fields)
-    }
-  })
-}
-const message = (mes: any, mesType: any): void => {
-  //@ts-ignore
-  ElMessage({
-    message: mes,
-    type: mesType,
-    duration: 1500
-  })
-}
-const fetchSignin = async (data: SigninInputType) => {
-  try {
-    isLoading.value = true
-    const response: any = await authStore.signin(data)
-    message(response.message, 'success')
-    router.push('/')
-  } catch (error: any) {
-    message(error.message, 'error')
-  } finally {
-    isLoading.value = false
-  }
-}
+// const isLoading = ref<boolean>(false)
+// type SigninInputType = {
+//   account: string
+//   password: string
+// }
+// const router: Router = useRouter()
+// const signinInput = ref<SigninInputType>({
+//   account: '',
+//   password: ''
+// })
+// const registerRules = ref({
+//   account: [
+//     {
+//       type: 'email',
+//       required: true,
+//       message: '信箱格式不相符',
+//       trigger: ['blur', 'change']
+//     }
+//   ],
+//   password: [
+//     { min: 2, max: 30, message: '長度介於6到30之間', trigger: 'blur' },
+//     { required: true, message: '必填', trigger: 'blur' }
+//   ]
+// })
+// const handleRegister = async (formEl: FormInstance | undefined) => {
+//   if (!formEl) return
+//   await formEl.validate(async (valid, fields) => {
+//     if (valid) {
+//       await fetchSignin(signinInput.value)
+//       // console.log('發送登入API')
+//     } else {
+//       // console.log('error submit!', fields)
+//     }
+//   })
+// }
+// const message = (mes: any, mesType: any): void => {
+//   //@ts-ignore
+//   ElMessage({
+//     message: mes,
+//     type: mesType,
+//     duration: 1500
+//   })
+// }
+// const fetchSignin = async (data: SigninInputType) => {
+//   try {
+//     isLoading.value = true
+//     const response: any = await authStore.signin(data)
+//     message(response.message, 'success')
+//     router.push('/')
+//   } catch (error: any) {
+//     message(error.message, 'error')
+//   } finally {
+//     isLoading.value = false
+//   }
+// }
+const signinInput = defineModel('signinInput', { type: Object })
+const props = defineProps<{
+  rules: FormRules
+  loading: boolean
+  handleSignin: (formEl: FormInstance | undefined) => Promise<void>
+  // formRef: FormInstance | undefined
+}>()
 </script>
 <template>
   <el-form
     class="el-form-font-size"
     ref="ruleFormRef"
-    :rules="registerRules"
+    :rules="rules"
     :model="signinInput"
-    v-loading="isLoading"
+    v-loading="loading"
   >
     <el-form-item label="登入電子信箱帳號:" label-position="top" prop="account">
       <el-input v-model="signinInput.account" />
@@ -80,7 +87,7 @@ const fetchSignin = async (data: SigninInputType) => {
     </div>
     <button
       class="w-full rounded-md bg-primary-300 px-4 py-3 text-center font-bold"
-      @click.prevent="handleRegister(ruleFormRef)"
+      @click.prevent="handleSignin(ruleFormRef)"
     >
       登入
     </button>
